+++ google2/CHANGELOG.md
@@ -10,6 +10,7 @@ Add
 
 Changed
 - Remove `retryConfig` from `PublisherConfig`
+- Update Kubernetes client library
 
 Dependency Upgrades
 ```
@@ -28,7 +29,7 @@ Update guava to 30.0-jre (#390)
 Update `io.kubernetes client-java` from `5.0.0` to `10.0.0` (This has some breaking changes if you're using the library's API directly)
 ```
       
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.17-c4e656f"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.17-TRAVIS-REPLACE-ME"`
 
 ## 0.16